Virgin Vue in vanilla HTML gone, chad webpack React introduced!
Vue is coold though. Anyway, lots of UI progress and it's all extremely promising. Keep note that the url for the app (in the QML file) will be changed, as localhost:8080 is only being used by React for development. It will be replace by the url to the HTML, built by webpack.